# docs/api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/screenings":{"get":{"tags":["screening-controller"],"operationId":"readMovieScreeningsInCinemaByStartAndEndDate","parameters":[{"name":"movieId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"cinemaId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"startDate","in":"query","required":true,"schema":{"type":"string"}},{"name":"endDate","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ScreeningResponse"}}}}}}},"post":{"tags":["screening-controller"],"operationId":"createScreening","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScreeningRequest"}}},"required":true},"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ScreeningResponse"}}}}}}},"/reservations":{"post":{"tags":["reservation-controller"],"operationId":"createReservation","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReservationRequest"}}},"required":true},"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReservationResponse"}}}}}}},"/reservationPrice":{"post":{"tags":["reservation-controller"],"operationId":"calculateReservationPrice","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReservationPriceRequest"}}},"required":true},"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReservationPriceResponse"}}}}}}},"/movies":{"get":{"tags":["movie-controller"],"operationId":"getMovies","responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MovieResponse"}}}}}}},"post":{"tags":["movie-controller"],"operationId":"createMovie","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MovieRequest"}}},"required":true},"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MovieResponse"}}}}}}},"/api/user":{"post":{"tags":["user-with-role-controller"],"summary":"Add a new UserWithRoles user","description":"If a default role is defined (app.default-role ), this role will be assigned to the user.","operationId":"addUserWithRoles","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserWithRolesRequest"}}},"required":true},"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserWithRolesResponse"}}}}}}},"/api/auth/login":{"post":{"tags":["authentication-controller"],"summary":"Login","description":"Use this to login and get a token","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}}}}},"/api/user/remove-role/{username}/{role}":{"patch":{"tags":["user-with-role-controller"],"summary":"Removes a role from a UserWithRoles","description":"Caller must be authenticated with the role ADMIN","operationId":"removeRole","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string"}},{"name":"role","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserWithRolesResponse"}}}}}}},"/api/user/add-role/{username}/{role}":{"patch":{"tags":["user-with-role-controller"],"summary":"Add a role to a UserWithRoles","description":"Caller must be authenticated with the role ADMIN","operationId":"addRole","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string"}},{"name":"role","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserWithRolesResponse"}}}}}}},"/seats/screening/{screeningId}":{"get":{"tags":["seat-controller"],"operationId":"getReservedSeatsByScreeningId","parameters":[{"name":"screeningId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SeatResponse"}}}}}}}},"/seats/auditorium/{auditoriumId}":{"get":{"tags":["seat-controller"],"operationId":"getSeatsFromAuditorium","parameters":[{"name":"auditoriumId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SeatResponse"}}}}}}}},"/reservations/users/{userId}":{"get":{"tags":["reservation-controller"],"operationId":"getReservationsByUserId","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ReservationResponse"}}}}}}}},"/priceadjustments":{"get":{"tags":["price-adjustment-controller"],"operationId":"getPriceAdjustments","responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PriceAdjustmentResponse"}}}}}}},"/movies/cinemas/{cinema}":{"get":{"tags":["movie-controller"],"operationId":"getMoviesByCinema","parameters":[{"name":"cinema","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MovieResponse"}}}}}}}},"/movies/TMDB/{id}":{"get":{"tags":["movie-controller"],"operationId":"getTMDBMovie","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MovieDetailsResponse"}}}}}}},"/cinemas":{"get":{"tags":["cinema-controller"],"operationId":"getCinemas","responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CinemaResponse"}}}}}}}},"/cinemas/{id}":{"get":{"tags":["cinema-controller"],"operationId":"getCinema","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CinemaResponse"}}}}}}},"/cinemas/{cinemaId}/auditoriums":{"get":{"tags":["cinema-controller"],"operationId":"getAllAuditoriumsByCinemaId","parameters":[{"name":"cinemaId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AuditoriumResponse"}}}}}}}}},"components":{"schemas":{"ScreeningRequest":{"type":"object","properties":{"movieId":{"type":"integer","format":"int64"},"auditoriumId":{"type":"integer","format":"int64"},"startTime":{"type":"string","format":"date-time"},"is3D":{"type":"boolean"}}},"AuditoriumResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"}}},"MovieResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"poster":{"type":"string"},"runtime":{"type":"integer","format":"int32"},"premiere":{"type":"string","format":"date"}}},"ScreeningResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"auditorium":{"$ref":"#/components/schemas/AuditoriumResponse"},"movie":{"$ref":"#/components/schemas/MovieResponse"},"startTime":{"type":"string","format":"date-time"},"is3D":{"type":"boolean"}}},"ReservationRequest":{"type":"object","properties":{"screeningId":{"type":"integer","format":"int64"},"seatIds":{"type":"array","items":{"type":"integer","format":"int64"}}}},"ReservationResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"user":{"type":"string"},"screening":{"$ref":"#/components/schemas/ScreeningResponse"},"seats":{"type":"array","items":{"$ref":"#/components/schemas/SeatResponse"}},"createdAt":{"type":"string","format":"date-time"}}},"SeatPricing":{"required":["price"],"type":"object","properties":{"name":{"type":"string"},"price":{"type":"number","format":"double"}}},"SeatResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"seatNumber":{"type":"integer","format":"int32"},"rowNumber":{"type":"integer","format":"int32"},"auditoriumId":{"type":"integer","format":"int64"},"seatPricing":{"$ref":"#/components/schemas/SeatPricing"}}},"ReservationPriceRequest":{"type":"object","properties":{"seatIds":{"type":"array","items":{"type":"integer","format":"int64"}},"screeningId":{"type":"integer","format":"int64"}}},"ReservationPriceResponse":{"type":"object","properties":{"seats":{"type":"number","format":"double"},"fees":{"type":"number","format":"double"},"discounts":{"type":"number","format":"double"},"totalPrice":{"type":"number","format":"double"}}},"MovieRequest":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"runtime":{"type":"integer","format":"int32"},"premiere":{"type":"string","format":"date"},"poster":{"type":"string"}}},"UserWithRolesRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"email":{"type":"string"}}},"UserWithRolesResponse":{"type":"object","properties":{"userName":{"type":"string"},"roleNames":{"type":"array","items":{"type":"string"}},"email":{"type":"string"}}},"LoginRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"LoginResponse":{"type":"object","properties":{"username":{"type":"string"},"token":{"type":"string"},"roles":{"type":"array","items":{"type":"string"}},"email":{"type":"string"}}},"PriceAdjustmentResponse":{"type":"object","properties":{"priceAdjustments":{"type":"object","additionalProperties":{"type":"number","format":"double"}}}},"MovieDetailsResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"title":{"type":"string"},"overview":{"type":"string"},"runtime":{"type":"integer","format":"int32"},"status":{"type":"string"},"posterPath":{"type":"string"},"genres":{"type":"array","items":{"type":"string"}}}},"CinemaResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"city":{"type":"string"},"isActive":{"type":"boolean"}}}}}}